refactor(philips-hue): Syntax fixes
